refactor(app_runtime): Profile handler を profile.rs に抽出 (SPEC-2077 Phase E redirect)
- Phase E は当初 Agent handler だったが、 mod.rs に純粋な agent _events 関数が無いため Profile handler 抽出に redirect
- crates/gwt/src/app_runtime/profile.rs を新規作成し、 6 profile handler (load / select / create / set_active / save / delete) と private helper (resolve_profile_window_context / profile_window_ids_for_tab / profile_config_path / active_profile_spawn_env / profile_snapshot_events) と ProfileSaveRequest payload を移行
- mod.rs に mod profile; 宣言 + use profile::ProfileSaveRequest を追加
- god module 解体 phase chain Phase E、 dispatch 経路 0 変更で blast radius 最小

Verification:
- cargo test -p gwt-core -p gwt: 全 pass
- cargo clippy -- -D warnings: 0 warnings
- cargo fmt --check: 差分なし
- mod.rs 6865→6505 行 (Phase E で 360 行減、 累計 940 行減)
- profile.rs 新 395 行 (Profile handler 単一責務)
