+++ crates/gwt/src/app_runtime/mod.rs
@@ -145,8 +145,10 @@ pub(crate) type AgentLaunchResult = Result<AgentLaunchCompletion, String>;
 mod board;
 mod memory;
 mod migration;
+mod profile;
 mod window;
 pub(crate) use board::BoardPostRequest;
+use profile::ProfileSaveRequest;
 
 fn dispatch_agent_launch_success<F>(
     proxy: AppEventProxy,
@@ -352,14 +354,6 @@ fn knowledge_view_events(
             },
         ),
     ]
-}
-
-struct ProfileSaveRequest {
-    current_name: String,
-    name: String,
-    description: String,
-    env_vars: Vec<gwt::ProfileEnvEntryView>,
-    disabled_env: Vec<String>,
 }
 
 pub(crate) fn build_frontend_sync_events(
@@ -1340,52 +1334,6 @@ impl AppRuntime {
         }
     }
 
-    pub(crate) fn load_profile_events(&mut self, client_id: &str, id: &str) -> Vec<OutboundEvent> {
-        if let Err(message) = self.resolve_profile_window_context(id) {
-            return vec![OutboundEvent::reply(
-                client_id,
-                BackendEvent::ProfileError {
-                    id: id.to_string(),
-                    message,
-                },
-            )];
-        }
-
-        let selected_profile = self.profile_selections.get(id).cloned();
-        let config_path = match self.profile_config_path() {
-            Ok(path) => path,
-            Err(message) => {
-                return vec![OutboundEvent::reply(
-                    client_id,
-                    BackendEvent::ProfileError {
-                        id: id.to_string(),
-                        message,
-                    },
-                )];
-            }
-        };
-        match gwt::profile_dispatch::load_snapshot_at(&config_path, selected_profile.as_deref()) {
-            Ok(snapshot) => {
-                self.profile_selections
-                    .insert(id.to_string(), snapshot.selected_profile.clone());
-                vec![OutboundEvent::reply(
-                    client_id,
-                    BackendEvent::ProfileSnapshot {
-                        id: id.to_string(),
-                        snapshot,
-                    },
-                )]
-            }
-            Err(error) => vec![OutboundEvent::reply(
-                client_id,
-                BackendEvent::ProfileError {
-                    id: id.to_string(),
-                    message: error.to_string(),
-                },
-            )],
-        }
-    }
-
     pub(crate) fn load_logs_events(&self, client_id: &str, id: &str) -> Vec<OutboundEvent> {
         let Some(address) = self.window_lookup.get(id) else {
             return vec![OutboundEvent::reply(
@@ -1440,314 +1388,6 @@ impl AppRuntime {
                 },
             )],
         }
-    }
-
-    pub(crate) fn select_profile_events(
-        &mut self,
-        client_id: &str,
-        id: &str,
-        profile_name: &str,
-    ) -> Vec<OutboundEvent> {
-        if let Err(message) = self.resolve_profile_window_context(id) {
-            return vec![OutboundEvent::reply(
-                client_id,
-                BackendEvent::ProfileError {
-                    id: id.to_string(),
-                    message,
-                },
-            )];
-        }
-
-        self.profile_selections
-            .insert(id.to_string(), profile_name.to_string());
-        self.load_profile_events(client_id, id)
-    }
-
-    pub(crate) fn create_profile_events(
-        &mut self,
-        client_id: &str,
-        id: &str,
-        name: &str,
-    ) -> Vec<OutboundEvent> {
-        let tab_id = match self.resolve_profile_window_context(id) {
-            Ok(tab_id) => tab_id,
-            Err(message) => {
-                return vec![OutboundEvent::reply(
-                    client_id,
-                    BackendEvent::ProfileError {
-                        id: id.to_string(),
-                        message,
-                    },
-                )];
-            }
-        };
-
-        let selected_profile = name.trim().to_string();
-        let config_path = match self.profile_config_path() {
-            Ok(path) => path,
-            Err(message) => {
-                return vec![OutboundEvent::reply(
-                    client_id,
-                    BackendEvent::ProfileError {
-                        id: id.to_string(),
-                        message,
-                    },
-                )];
-            }
-        };
-        if let Err(error) =
-            gwt::profile_dispatch::create_profile_at(&config_path, &selected_profile)
-        {
-            return vec![OutboundEvent::reply(
-                client_id,
-                BackendEvent::ProfileError {
-                    id: id.to_string(),
-                    message: error.to_string(),
-                },
-            )];
-        }
-
-        self.profile_selections
-            .insert(id.to_string(), selected_profile);
-        self.profile_snapshot_events(&tab_id, id, client_id)
-    }
-
-    pub(crate) fn set_active_profile_events(
-        &mut self,
-        client_id: &str,
-        id: &str,
-        profile_name: &str,
-    ) -> Vec<OutboundEvent> {
-        let tab_id = match self.resolve_profile_window_context(id) {
-            Ok(tab_id) => tab_id,
-            Err(message) => {
-                return vec![OutboundEvent::reply(
-                    client_id,
-                    BackendEvent::ProfileError {
-                        id: id.to_string(),
-                        message,
-                    },
-                )];
-            }
-        };
-
-        let config_path = match self.profile_config_path() {
-            Ok(path) => path,
-            Err(message) => {
-                return vec![OutboundEvent::reply(
-                    client_id,
-                    BackendEvent::ProfileError {
-                        id: id.to_string(),
-                        message,
-                    },
-                )];
-            }
-        };
-        if let Err(error) =
-            gwt::profile_dispatch::switch_active_profile_at(&config_path, profile_name)
-        {
-            return vec![OutboundEvent::reply(
-                client_id,
-                BackendEvent::ProfileError {
-                    id: id.to_string(),
-                    message: error.to_string(),
-                },
-            )];
-        }
-
-        self.profile_selections
-            .insert(id.to_string(), profile_name.to_string());
-        self.profile_snapshot_events(&tab_id, id, client_id)
-    }
-
-    fn save_profile_events(
-        &mut self,
-        client_id: &str,
-        id: &str,
-        request: ProfileSaveRequest,
-    ) -> Vec<OutboundEvent> {
-        let tab_id = match self.resolve_profile_window_context(id) {
-            Ok(tab_id) => tab_id,
-            Err(message) => {
-                return vec![OutboundEvent::reply(
-                    client_id,
-                    BackendEvent::ProfileError {
-                        id: id.to_string(),
-                        message,
-                    },
-                )];
-            }
-        };
-
-        let config_path = match self.profile_config_path() {
-            Ok(path) => path,
-            Err(message) => {
-                return vec![OutboundEvent::reply(
-                    client_id,
-                    BackendEvent::ProfileError {
-                        id: id.to_string(),
-                        message,
-                    },
-                )];
-            }
-        };
-        if let Err(error) = gwt::profile_dispatch::save_profile_at(
-            &config_path,
-            &request.current_name,
-            &request.name,
-            &request.description,
-            &request.env_vars,
-            &request.disabled_env,
-        ) {
-            return vec![OutboundEvent::reply(
-                client_id,
-                BackendEvent::ProfileError {
-                    id: id.to_string(),
-                    message: error.to_string(),
-                },
-            )];
-        }
-
-        self.profile_selections
-            .insert(id.to_string(), request.name.trim().to_string());
-        self.profile_snapshot_events(&tab_id, id, client_id)
-    }
-
-    pub(crate) fn delete_profile_events(
-        &mut self,
-        client_id: &str,
-        id: &str,
-        profile_name: &str,
-    ) -> Vec<OutboundEvent> {
-        let tab_id = match self.resolve_profile_window_context(id) {
-            Ok(tab_id) => tab_id,
-            Err(message) => {
-                return vec![OutboundEvent::reply(
-                    client_id,
-                    BackendEvent::ProfileError {
-                        id: id.to_string(),
-                        message,
-                    },
-                )];
-            }
-        };
-
-        let config_path = match self.profile_config_path() {
-            Ok(path) => path,
-            Err(message) => {
-                return vec![OutboundEvent::reply(
-                    client_id,
-                    BackendEvent::ProfileError {
-                        id: id.to_string(),
-                        message,
-                    },
-                )];
-            }
-        };
-        if let Err(error) = gwt::profile_dispatch::delete_profile_at(&config_path, profile_name) {
-            return vec![OutboundEvent::reply(
-                client_id,
-                BackendEvent::ProfileError {
-                    id: id.to_string(),
-                    message: error.to_string(),
-                },
-            )];
-        }
-
-        self.profile_snapshot_events(&tab_id, id, client_id)
-    }
-
-    fn resolve_profile_window_context(&self, id: &str) -> std::result::Result<String, String> {
-        let Some(address) = self.window_lookup.get(id) else {
-            return Err("Window not found".to_string());
-        };
-        let Some(tab) = self.tab(&address.tab_id) else {
-            return Err("Project tab not found".to_string());
-        };
-        let Some(window) = tab.workspace.window(&address.raw_id) else {
-            return Err("Window not found".to_string());
-        };
-        if window.preset != WindowPreset::Profile {
-            return Err("Window is not a Profile surface".to_string());
-        }
-
-        Ok(address.tab_id.clone())
-    }
-
-    fn profile_window_ids_for_tab(&self, tab_id: &str) -> Vec<String> {
-        let Some(tab) = self.tab(tab_id) else {
-            return Vec::new();
-        };
-        tab.workspace
-            .persisted()
-            .windows
-            .iter()
-            .filter(|window| window.preset == WindowPreset::Profile)
-            .map(|window| combined_window_id(tab_id, &window.id))
-            .collect()
-    }
-
-    fn profile_config_path(&self) -> std::result::Result<PathBuf, String> {
-        if let Some(path) = &self.profile_config_path {
-            return Ok(path.clone());
-        }
-        gwt::profile_dispatch::config_path().map_err(|error| error.to_string())
-    }
-
-    fn active_profile_spawn_env(&self) -> Result<gwt_agent::LaunchEnvironment, String> {
-        let config_path = self.profile_config_path()?;
-        gwt_agent::LaunchEnvironment::from_active_profile(
-            &config_path,
-            gwt_agent::LaunchRuntimeTarget::Host,
-        )
-    }
-
-    fn profile_snapshot_events(
-        &mut self,
-        tab_id: &str,
-        selected_window_id: &str,
-        client_id: &str,
-    ) -> Vec<OutboundEvent> {
-        let window_ids = self.profile_window_ids_for_tab(tab_id);
-        let mut events = Vec::new();
-
-        for window_id in window_ids {
-            let selected_profile = self.profile_selections.get(&window_id).cloned();
-            let config_path = match self.profile_config_path() {
-                Ok(path) => path,
-                Err(message) => {
-                    return vec![OutboundEvent::reply(
-                        client_id,
-                        BackendEvent::ProfileError {
-                            id: selected_window_id.to_string(),
-                            message,
-                        },
-                    )];
-                }
-            };
-            match gwt::profile_dispatch::load_snapshot_at(&config_path, selected_profile.as_deref())
-            {
-                Ok(snapshot) => {
-                    self.profile_selections
-                        .insert(window_id.clone(), snapshot.selected_profile.clone());
-                    events.push(OutboundEvent::broadcast(BackendEvent::ProfileSnapshot {
-                        id: window_id,
-                        snapshot,
-                    }));
-                }
-                Err(error) => {
-                    return vec![OutboundEvent::reply(
-                        client_id,
-                        BackendEvent::ProfileError {
-                            id: selected_window_id.to_string(),
-                            message: error.to_string(),
-                        },
-                    )];
-                }
-            }
-        }
-
-        events
     }
 
     pub(crate) fn handle_board_projection_changed_events(
